Re-organize source structures in SearchControls.
