+++ components/webui/imports/ui/SearchView/SearchControls/SearchControlsFilterDrawer/SearchControlsFilterDrawer.scss
@@ -1,0 +1,12 @@
+.search-filter {
+  &-controls-drawer {
+    background-color: #efefef;
+  }
+
+  &-control-label {
+    color: #666;
+
+    min-width: 9rem;
+  }
+}
+
