+++ components/webui/imports/ui/SearchView/SearchControls/SearchControlsFilterDrawer/SearchControlsTimeRangeInput/SearchControlsTimeRangeInput.scss
@@ -1,15 +1,3 @@
-.search-filter {
-  &-controls-drawer {
-    background-color: #efefef;
-  }
-
-  &-control-label {
-    color: #666;
-
-    min-width: 9rem;
-  }
-}
-
 .timestamp-picker {
   border: 1px solid rgb(206, 212, 218);
   border-radius: 0;
